+++ site/gatsby-site/src/components/RelatedIncidentsArea.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import { ListGroup, Card, Spinner, Button, ButtonGroup, ButtonToolbar } from 'react-bootstrap';
+import { ListGroup, Card, Spinner, Button, ButtonToolbar } from 'react-bootstrap';
 import { useFormikContext } from 'formik';
+import SimilaritySelector from './SimilaritySelector';
 
 const ListContainer = styled(Card)`
   margin: 1em 0;
@@ -30,22 +31,8 @@ const ReportRow = styled(ListGroup.Item)`
   }
 `;
 
-const SimilarSelector = styled(ButtonGroup)`
+const SimilarSelector = styled(SimilaritySelector)`
   margin-left: 1ch;
-`;
-
-const SimilarityButton = styled(Button)`
-  transition: 0.2s ease-in all !important;
-  opacity: 0.8 !important;
-  width: 2.5em;
-  padding-left: 0px !important;
-  padding-right: 0px !important;
-  text-align: center;
-
-  :disabled {
-    opacity: 1 !important;
-    box-shadow: inset 0px 0px 4px 4px rgba(0, 0, 0, 0.05);
-  }
 `;
 
 const ReportToolbar = styled(ButtonToolbar)`
@@ -66,7 +53,7 @@ const RelatedIncidentsArea = ({
     return null;
   }
 
-  const { values, setFieldValue } = editable ? useFormikContext() : { setFieldValue: null };
+  const { setFieldValue } = editable ? useFormikContext() : { setFieldValue: null };
 
   return (
     <ListContainer data-cy={`related-${columnKey}`}>
@@ -92,85 +79,7 @@ const RelatedIncidentsArea = ({
                 {editable && (
                   <>
                     <label htmlFor="similar-selector">Related: </label>
-                    <SimilarSelector data-cy="similar-selector" id="similar-selector">
-                      {[
-                        {
-                          identifier: 'dissimilar',
-                          variant: 'danger',
-                          icon: 'x',
-                          show:
-                            values.editor_dissimilar_incidents &&
-                            values.editor_dissimilar_incidents.includes(val.incident_id),
-                          onClick: () => {
-                            setFieldValue(
-                              'editor_dissimilar_incidents',
-                              (values.editor_dissimilar_incidents || []).concat([val.incident_id])
-                            );
-                            setFieldValue(
-                              'editor_similar_incidents',
-                              (values.editor_similar_incidents || []).filter(
-                                (id) => id != val.incident_id
-                              )
-                            );
-                          },
-                        },
-                        {
-                          identifier: 'unspecified',
-                          variant: 'secondary',
-                          icon: '?',
-                          show:
-                            !values.editor_similar_incidents ||
-                            !values.editor_dissimilar_incidents ||
-                            (!values.editor_similar_incidents.includes(val.incident_id) &&
-                              !values.editor_dissimilar_incidents.includes(val.incident_id)),
-                          onClick: () => {
-                            setFieldValue(
-                              'editor_similar_incidents',
-                              (values.editor_similar_incidents || []).filter(
-                                (id) => id != val.incident_id
-                              )
-                            );
-                            setFieldValue(
-                              'editor_dissimilar_incidents',
-                              (values.editor_dissimilar_incidents || []).filter(
-                                (id) => id != val.incident_id
-                              )
-                            );
-                          },
-                        },
-                        {
-                          identifier: 'similar',
-                          variant: 'success',
-                          icon: '✓',
-                          show:
-                            values.editor_similar_incidents &&
-                            values.editor_similar_incidents.includes(val.incident_id),
-                          onClick: () => {
-                            setFieldValue(
-                              'editor_similar_incidents',
-                              (values.editor_similar_incidents || []).concat([val.incident_id])
-                            );
-                            setFieldValue(
-                              'editor_dissimilar_incidents',
-                              (values.editor_dissimilar_incidents || []).filter(
-                                (id) => id != val.incident_id
-                              )
-                            );
-                          },
-                        },
-                      ].map((button) => (
-                        <SimilarityButton
-                          variant={button.show ? button.variant : 'secondary'}
-                          aria-pressed={button.show}
-                          disabled={button.show}
-                          onClick={button.onClick}
-                          key={button.icon}
-                          data-cy={button.identifier}
-                        >
-                          {button.icon}
-                        </SimilarityButton>
-                      ))}
-                    </SimilarSelector>
+                    <SimilarSelector incident_id={val.incident_id} />
                   </>
                 )}
                 {val.incident_id && editable && editId && (
